backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
httpx==0.28.1
apscheduler==3.10.4
websockets==14.1
